# Instantiate proxy classes via reflection, remove Enhancer.createFactory