Here's alternative approach to class instantiation (this PR is alternative to #79).

Note: the same generated class might be used with different constructors later, thus there should be something like an inline cache (otherwise it will end up with some kind of search for the proper constructor).

Here I implement 1-entry inline cache (== remember just one constructor for the class, and use it).
If user uses more than one constructor per proxy, the code falls back to `ReflectionUtils.newInstance(...)`, and it will be rather slow: https://github.com/cglib/cglib/pull/82/files#diff-173a9c4b512f39086ed564b6320318e2R422

The results are as follows:

# Before

```
1.7u55, factory
Benchmark                                                    Mode  Cnt     Score    Error   Units
BeansBenchmark.newInstance                                   avgt   10   201,222 ±  3,494   ns/op
BeansBenchmark.newInstance:·gc.alloc.rate                    avgt   10  1251,090 ± 21,714  MB/sec
BeansBenchmark.newInstance:·gc.alloc.rate.norm               avgt   10   264,000 ±  0,001    B/op

1.8u60, factory
Benchmark                                                    Mode  Cnt     Score    Error   Units
BeansBenchmark.newInstance                                   avgt   10   146,819 ±  2,805   ns/op
BeansBenchmark.newInstance:·gc.alloc.rate.norm               avgt   10   264,000 ±  0,001    B/op
```


# After

NOTE: the results are for happy-case scenario (i.e. just one constructor is used)
Otherwise the code falls back to old `getDeclaredConstructor` path: https://github.com/cglib/cglib/pull/82/files#diff-173a9c4b512f39086ed564b6320318e2R422

```
1.7u55, reflection
Benchmark                                                    Mode  Cnt     Score    Error   Units
BeansBenchmark.newInstance                                   avgt   10   208,021 ±  6,498   ns/op
BeansBenchmark.newInstance:·gc.alloc.rate.norm               avgt   10   256,000 ±  0,001    B/op

1.8u60, reflection
Benchmark                                                    Mode  Cnt     Score    Error   Units
BeansBenchmark.newInstance                                   avgt   10   162,185 ±  3,051   ns/op
BeansBenchmark.newInstance:·gc.alloc.rate.norm               avgt   10   256,000 ±  0,001    B/op
```

closes #75