@sameb 2016-06-23
lgtm with the exception of some comment clarifications.  LMK if you're able to update the comments and i'll merge after.  thanks!