@sameb 2016-06-23
Could you add a test that this fixes #75 ?